Emits different software fault events. Remove optional events that aren't emited by the app
