+++ examples/platform/silabs/silabs_utils.cpp
@@ -18,15 +18,18 @@
  */
 
 #include "silabs_utils.h"
-
+#include "SoftwareFaultReports.h"
 #include <matter_config.h>
 
 void appError(int err)
 {
-    SILABS_LOG("!!!!!!!!!!!! App Critical Error: %d !!!!!!!!!!!", err);
-    portDISABLE_INTERRUPTS();
-    while (true)
-        ;
+    char faultMessage[30] = { 0 }; // App Critical Error: + 10 digit + null character
+    snprintf(faultMessage, sizeof faultMessage, "App Critical Error:%d", err);
+    SILABS_LOG("!!!!!!!!!!!! %s !!!!!!!!!!!", faultMessage);
+    chip::DeviceLayer::Silabs::OnSoftwareFaultEventHandler(faultMessage);
+    vTaskDelay(pdMS_TO_TICKS(1000));
+    /* Force an assert. */
+    chipAbort();
 }
 
 void appError(CHIP_ERROR error)
