+++ src/platform/silabs/Logging.cpp
@@ -378,191 +378,3 @@ extern "C" void otPlatLog(otLogLevel aLogLevel, otLogRegion aLogRegion, const ch
     va_end(v);
 }
 #endif // CHIP_ENABLE_OPENTHREAD
-
-#if HARD_FAULT_LOG_ENABLE && SILABS_LOG_ENABLED
-
-/**
- * Log register contents to UART when a hard fault occurs.
- */
-extern "C" void debugHardfault(uint32_t * sp)
-{
-    uint32_t cfsr  = SCB->CFSR;
-    uint32_t hfsr  = SCB->HFSR;
-    uint32_t mmfar = SCB->MMFAR;
-    uint32_t bfar  = SCB->BFAR;
-    uint32_t r0    = sp[0];
-    uint32_t r1    = sp[1];
-    uint32_t r2    = sp[2];
-    uint32_t r3    = sp[3];
-    uint32_t r12   = sp[4];
-    uint32_t lr    = sp[5];
-    uint32_t pc    = sp[6];
-    uint32_t psr   = sp[7];
-    char formattedMsg[32];
-
-    if (sLogInitialized == false)
-    {
-        silabsInitLog();
-    }
-
-    snprintf(formattedMsg, sizeof formattedMsg, LOG_ERROR "HardFault:\n");
-    PrintLog(formattedMsg);
-    snprintf(formattedMsg, sizeof formattedMsg, "SCB->CFSR   0x%08lx", cfsr);
-    PrintLog(formattedMsg);
-    snprintf(formattedMsg, sizeof formattedMsg, "SCB->HFSR   0x%08lx", hfsr);
-    PrintLog(formattedMsg);
-    snprintf(formattedMsg, sizeof formattedMsg, "SCB->MMFAR  0x%08lx", mmfar);
-    PrintLog(formattedMsg);
-    snprintf(formattedMsg, sizeof formattedMsg, "SCB->BFAR   0x%08lx", bfar);
-    PrintLog(formattedMsg);
-    snprintf(formattedMsg, sizeof formattedMsg, "SP          0x%08lx", (uint32_t) sp);
-    PrintLog(formattedMsg);
-    snprintf(formattedMsg, sizeof formattedMsg, "R0          0x%08lx\n", r0);
-    PrintLog(formattedMsg);
-    snprintf(formattedMsg, sizeof formattedMsg, "R1          0x%08lx\n", r1);
-    PrintLog(formattedMsg);
-    snprintf(formattedMsg, sizeof formattedMsg, "R2          0x%08lx\n", r2);
-    PrintLog(formattedMsg);
-    snprintf(formattedMsg, sizeof formattedMsg, "R3          0x%08lx\n", r3);
-    PrintLog(formattedMsg);
-    snprintf(formattedMsg, sizeof formattedMsg, "R12         0x%08lx\n", r12);
-    PrintLog(formattedMsg);
-    snprintf(formattedMsg, sizeof formattedMsg, "LR          0x%08lx\n", lr);
-    PrintLog(formattedMsg);
-    snprintf(formattedMsg, sizeof formattedMsg, "PC          0x%08lx\n", pc);
-    PrintLog(formattedMsg);
-    snprintf(formattedMsg, sizeof formattedMsg, "PSR         0x%08lx\n", psr);
-    PrintLog(formattedMsg);
-
-    while (true)
-        ;
-}
-
-/**
- * Override default hard-fault handler
- */
-extern "C" __attribute__((naked)) void HardFault_Handler(void)
-{
-    __asm volatile("tst lr, #4                                    \n"
-                   "ite eq                                        \n"
-                   "mrseq r0, msp                                 \n"
-                   "mrsne r0, psp                                 \n"
-                   "ldr r1, debugHardfault_address                \n"
-                   "bx r1                                         \n"
-                   "debugHardfault_address: .word debugHardfault  \n");
-}
-#endif // HARD_FAULT_LOG_ENABLE && SILABS_LOG_ENABLED
-
-#if HARD_FAULT_LOG_ENABLE
-extern "C" void vApplicationMallocFailedHook(void)
-{
-    /* Called if a call to pvPortMalloc() fails because there is insufficient
-    free memory available in the FreeRTOS heap.  pvPortMalloc() is called
-    internally by FreeRTOS API functions that create tasks, queues, software
-    timers, and semaphores.  The size of the FreeRTOS heap is set by the
-    configTOTAL_HEAP_SIZE configuration constant in FreeRTOSConfig.h. */
-
-#if SILABS_LOG_ENABLED
-    SILABS_LOG("Failed do a malloc on HEAP. Is it too small ?");
-#endif
-    /* Force an assert. */
-    configASSERT((volatile void *) NULL);
-}
-/*-----------------------------------------------------------*/
-
-extern "C" void vApplicationStackOverflowHook(TaskHandle_t pxTask, char * pcTaskName)
-{
-    //(void) pcTaskName;
-    (void) pxTask;
-
-    /* Run time stack overflow checking is performed if
-    configCHECK_FOR_STACK_OVERFLOW is defined to 1 or 2.  This hook
-    function is called if a stack overflow is detected. */
-
-#if SILABS_LOG_ENABLED
-    SILABS_LOG("TASK OVERFLOW");
-    SILABS_LOG(pcTaskName);
-#endif
-    /* Force an assert. */
-    configASSERT((volatile void *) NULL);
-}
-
-extern "C" void vApplicationTickHook(void) {}
-
-/*-----------------------------------------------------------*/
-
-/* configUSE_STATIC_ALLOCATION is set to 1, so the application must provide an
-implementation of vApplicationGetIdleTaskMemory() to provide the memory that is
-used by the Idle task. */
-extern "C" void vApplicationGetIdleTaskMemory(StaticTask_t ** ppxIdleTaskTCBBuffer, StackType_t ** ppxIdleTaskStackBuffer,
-                                              uint32_t * pulIdleTaskStackSize)
-{
-    /* If the buffers to be provided to the Idle task are declared inside this
-    function then they must be declared static - otherwise they will be allocated on
-    the stack and so not exists after this function exits. */
-    static StaticTask_t xIdleTaskTCB;
-    static StackType_t uxIdleTaskStack[configMINIMAL_STACK_SIZE];
-
-    /* Pass out a pointer to the StaticTask_t structure in which the Idle task's
-    state will be stored. */
-    *ppxIdleTaskTCBBuffer = &xIdleTaskTCB;
-
-    /* Pass out the array that will be used as the Idle task's stack. */
-    *ppxIdleTaskStackBuffer = uxIdleTaskStack;
-
-    /* Pass out the size of the array pointed to by *ppxIdleTaskStackBuffer.
-    Note that, as the array is necessarily of type StackType_t,
-    configMINIMAL_STACK_SIZE is specified in words, not bytes. */
-    *pulIdleTaskStackSize = configMINIMAL_STACK_SIZE;
-}
-/*-----------------------------------------------------------*/
-
-/* configUSE_STATIC_ALLOCATION and configUSE_TIMERS are both set to 1, so the
-application must provide an implementation of vApplicationGetTimerTaskMemory()
-to provide the memory that is used by the Timer service task. */
-extern "C" void vApplicationGetTimerTaskMemory(StaticTask_t ** ppxTimerTaskTCBBuffer, StackType_t ** ppxTimerTaskStackBuffer,
-                                               uint32_t * pulTimerTaskStackSize)
-{
-    /* If the buffers to be provided to the Timer task are declared inside this
-    function then they must be declared static - otherwise they will be allocated on
-    the stack and so not exists after this function exits. */
-    static StaticTask_t xTimerTaskTCB;
-    static StackType_t uxTimerTaskStack[configTIMER_TASK_STACK_DEPTH];
-
-    /* Pass out a pointer to the StaticTask_t structure in which the Timer
-    task's state will be stored. */
-    *ppxTimerTaskTCBBuffer = &xTimerTaskTCB;
-
-    /* Pass out the array that will be used as the Timer task's stack. */
-    *ppxTimerTaskStackBuffer = uxTimerTaskStack;
-
-    /* Pass out the size of the array pointed to by *ppxTimerTaskStackBuffer.
-    Note that, as the array is necessarily of type StackType_t,
-    configMINIMAL_STACK_SIZE is specified in words, not bytes. */
-    *pulTimerTaskStackSize = configTIMER_TASK_STACK_DEPTH;
-}
-
-#ifndef BRD4325A
-extern "C" void RAILCb_AssertFailed(RAIL_Handle_t railHandle, uint32_t errorCode)
-{
-#if SILABS_LOG_ENABLED
-#ifdef RAIL_ASSERT_DEBUG_STRING
-    static const char * railErrorMessages[] = RAIL_ASSERT_ERROR_MESSAGES;
-    const char * errorMessage               = "Unknown";
-
-    If this error code is within the range of known error messages then use the appropriate error
-        message.if (errorCode < (sizeof(railErrorMessages) / sizeof(char *)))
-    {
-        errorMessage = railErrorMessages[errorCode];
-    }
-    SILABS_LOG("RAIL Assert : %s", errorMessage);
-#else
-    SILABS_LOG("RAIL Assert : %ld", errorCode);
-#endif // RAIL_ASSERT_DEBUG_STRING
-#endif // SILABS_LOG_ENABLED
-    while (1)
-        ;
-}
-#endif // BRD4325A
-
-#endif // HARD_FAULT_LOG_ENABLE
